cargo: Stop depending on crates from git repos
It makes it harder for libphosh-rs consumers to work with a local
checkout of this repo, as they might be depending on common crates (such
as gtk/glib) *not* from git which Cargo doesn't like very much.

Besides that, it's also preferable to build and test these crates
against the same dependencies that libphosh-rs consumers would be using
in most cases.

Signed-off-by: Sam Day <me@samcday.com>